+++ pmd-plsql/src/test/java/net/sourceforge/pmd/lang/plsql/ast/PlsqlTreeDumpTest.java
@@ -43,6 +43,11 @@ class PlsqlTreeDumpTest extends BaseTreeDumpTest {
     }
 
     @Test
+    void parseMergeStatement() {
+        doTest("MergeStatementIssue1934");
+    }
+  
+    @Test
     void errorLoggingClause() {
         doTest("ErrorLoggingClause2779");
     }